Move functions back into RoamImporter class
This helps reduce the number of args we need to pass around.
